// src/components/openapi/flow-apis/login-service/data-contracts.ts
export interface CreateTokenBody {
  /** The OAuth grant code */
  code: string;
  /** The state parameter */
  state: string;
}
// ...
export interface GeneratePrincipalPolicyUrlDto {
  /**
   * The email address of the user for whom the policy URL is being generated. Must be a valid email format.
   * @example "erikt@ciandt.com"
   */
  email: string;
  /**
   * An UUID code that will be used to store auth states further in the authentication flow.
   * @example "123e4567-e89b-12d3-a456-426614174000"
   */
  code: string;
  /**
   * The channel through which the request is made, affecting the resulting URL.
   * @example "vscode"
   */
  channel: string;
  /**
   * The redirect URI to which a user will be redirected after the authentication is completed.
   * @example "https://example.com/callback"
   */
  redirectUri: string;
}